MultichannelMixer ported to 64-bits, it crashes

+++ MultichannelMixer/MultichannelMixer/MultichannelMixerViewController.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace MultichannelMixer
 {
@@ -63,7 +63,7 @@ namespace MultichannelMixer
 		partial void setInputVolume (NSObject sender)
 		{
 			var slider = (UISlider) sender;
-			Mixer.SetInputVolume (slider.Tag, slider.Value);
+			Mixer.SetInputVolume ((int)slider.Tag, slider.Value);
 		}
 
 		partial void setOutputVolume (NSObject sender)
@@ -85,7 +85,7 @@ namespace MultichannelMixer
 				break;
 			}
 
-			Mixer.EnableInput (ctrl.Tag, ctrl.On);
+			Mixer.EnableInput ((int)ctrl.Tag, ctrl.On);
 		}
 
 		// called if we've been interrupted and if we're playing, stop
